refactor(settings): Remove missed file from jest coverage PR
