+++ packages/fxa-settings/src/components/ConnectedServices/index.stories.tsx
@@ -7,7 +7,7 @@ import { storiesOf } from '@storybook/react';
 import { LocationProvider } from '@reach/router';
 import { ConnectedServices } from '.';
 
-import { MOCK_SERVICES } from './MOCK_SERVICES';
+import { MOCK_SERVICES } from './mocks';
 import { AppContext } from 'fxa-settings/src/models';
 import { mockAppContext } from 'fxa-settings/src/models/mocks';
