+++ packages/fxa-settings/src/components/PageSettings/index.stories.tsx
@@ -9,7 +9,7 @@ import { LocationProvider } from '@reach/router';
 import AppLayout from '../AppLayout';
 import { isMobileDevice } from '../../lib/utilities';
 import { mockAppContext, mockEmail, MOCK_ACCOUNT } from '../../models/mocks';
-import { MOCK_SERVICES } from '../ConnectedServices/MOCK_SERVICES';
+import { MOCK_SERVICES } from '../ConnectedServices/mocks';
 import { AppContext } from 'fxa-settings/src/models';
 
 const SERVICES_NON_MOBILE = MOCK_SERVICES.filter((d) => !isMobileDevice(d));
